+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Change Log
 ==========
 
+## 2.1 alert_to_issue
+author: Andrew Wurster
+
+* Add support to create Jira issues based on Splunk alerts / saved searches.
+* Limited support for adding attributes (i.e. Summary, Desc, etc)
+* Setup Instructions
+    * Install using pip, where `--prefix=<SPLUNK_HOME>` so that Jira Python lib packages are available to Splunk
+        pip install --install-option="--prefix=/opt/splunk" jira-python
+    * link from Splunk scripts back to app
+        cd /opt/splunk/bin/scritps
+        ln -s  /opt/splunk/etc/apps/jira/bin/alert_to_jira.py alert_to_jira.py 
+
 ## 2.1
 
 * Add detail option to rapidboards mode